#167 JWTServiceTest divided into 2 test clases
