+++ api/src/test/scala/za/co/absa/loginsvc/rest/service/jwt/JWTServiceTest.scala
@@ -394,78 +394,4 @@ class JWTServiceTest extends AnyFlatSpec with BeforeAndAfterEach with Matchers {
     actualGroups shouldBe userWithManyGroups.groups
   }
 
-  behavior of "generateToken with allowProvidersToRefreshGroupsOnGenerate=true"
-
-  /**
-   * Replaces the default jwtService with one using allowProvidersToRefreshGroupsOnGenerate=true.
-   * afterEach will handle closing it, so no try/finally needed in individual tests.
-   */
-  private def useJwtServiceWithRefreshGroupsAllowed(): Unit = {
-    jwtService.close()
-    val configP = new JwtConfigProvider {
-      override def getJwtKeyConfig: KeyConfig = testConfig.getJwtKeyConfig match {
-        case inMemoryConfig: InMemoryKeyConfig =>
-          inMemoryConfig.copy(allowProvidersToRefreshGroupsOnGenerate = true)
-        case other => fail(s"Unexpected KeyConfig type: ${other.getClass.getName}")
-      }
-    }
-    jwtService = new JWTService(configP, authSearchService)
-  }
-
-  it should "fetch groups from UserSearchService when inputUser has empty groups and flag is true" in {
-    useJwtServiceWithRefreshGroupsAllowed()
-    // user2 exists in test config with groups = ["group2"]
-    val userWithEmptyGroups = User("user2", Seq.empty, Map("mail" -> Some("user@two.org")))
-    val jwt = jwtService.generateAccessToken(userWithEmptyGroups, None)
-    val parsedJWT = parseJWT(jwt)
-
-    assert(parsedJWT.isSuccess)
-    val actualGroups = parsedJWT.get.getBody.get("groups", classOf[util.ArrayList[String]]).asScala
-
-    actualGroups shouldBe Seq("group2")
-  }
-
-  Seq(
-    Some(PrefixesConfig(Set("gr"), caseSensitive = true)), // group2 should match
-    Some(PrefixesConfig(Set("GR"), caseSensitive = false)), // group2 should match case-insensitively
-  ).foreach { prefixConfig =>
-
-    it should s"fetch and filter groups from UserSearchService with prefix config when inputUser has empty groups (case $prefixConfig)" in {
-      useJwtServiceWithRefreshGroupsAllowed()
-      val userWithEmptyGroups = User("user2", Seq.empty, Map("mail" -> Some("user@two.org")))
-      val jwt = jwtService.generateAccessToken(userWithEmptyGroups, prefixConfig)
-      val parsedJWT = parseJWT(jwt)
-
-      assert(parsedJWT.isSuccess)
-      val actualGroups = parsedJWT.get.getBody.get("groups", classOf[util.ArrayList[String]]).asScala
-
-      actualGroups shouldBe Seq("group2")
-    }
-  }
-
-  it should "fetch and filter groups from UserSearchService resulting in empty when prefix doesn't match" in {
-    useJwtServiceWithRefreshGroupsAllowed()
-    val userWithEmptyGroups = User("user2", Seq.empty, Map("mail" -> Some("user@two.org")))
-    val prefixConfig = Some(PrefixesConfig(Set("nonexistent", "GR"), caseSensitive = true))
-    val jwt = jwtService.generateAccessToken(userWithEmptyGroups, prefixConfig)
-    val parsedJWT = parseJWT(jwt)
-
-    assert(parsedJWT.isSuccess)
-    val actualGroups = parsedJWT.get.getBody.get("groups", classOf[util.ArrayList[String]]).asScala
-
-    actualGroups shouldBe empty
-  }
-
-  it should "NOT fetch groups from UserSearchService when inputUser already has groups (flag is true but groups non-empty)" in {
-    useJwtServiceWithRefreshGroupsAllowed()
-    // user has groups already populated - should use them directly, not fetch from UserSearchService
-    val userWithExistingGroups = User("user2", Seq("custom-group"), Map("mail" -> Some("user@two.org")))
-    val jwt = jwtService.generateAccessToken(userWithExistingGroups, None)
-    val parsedJWT = parseJWT(jwt)
-
-    assert(parsedJWT.isSuccess)
-    val actualGroups = parsedJWT.get.getBody.get("groups", classOf[util.ArrayList[String]]).asScala
-
-    actualGroups shouldBe Seq("custom-group")
-  }
 }
